chore: Update NotifyUpdateInstallation component to use UpdateInstallationModal and remove NotifyUpdateInstallation
